[#noissue] Move the active thread socket into its own hook
Apply review: SonarQube flagged the component's cognitive complexity at
18 against a limit of 15. Opening the socket, parsing frames, sending
the request and backing off before a retry all lived inside the
component next to the rendering, and this PR pushed it over the line.

useActiveThreadSocket now owns the socket's whole life and hands back
the connection state and the counts. The component is left deciding
what to watch and drawing it. The moved code is unchanged.
